+++ main/core/commandline.c
@@ -331,6 +331,8 @@ void register_commands() {
     register_command("listenprobes", handle_listen_probes_cmd);
     register_command("settings", handle_settings_cmd);
     register_command("loglevel", handle_log_level_cmd);
+    register_command("fav", handle_fav_cmd);
+    register_command("favorites", handle_fav_cmd);
     register_command("listportals", handle_listportals);
     register_command("evilportal", handle_evilportal);
     register_command("commdiscovery", handle_comm_discovery);