+++ main/managers/views/options_screen.c
@@ -1,5 +1,6 @@
 #include "managers/views/options_screen.h"
 #include "managers/views/lockscreen.h"
+#include "managers/views/favorites_manager_screen.h"
 #include "core/serial_manager.h"
 #include "core/commandline.h"
 #include "core/ouis.h"
@@ -1066,6 +1067,7 @@ typedef enum {
     SETTINGS_CAT_DATE_TIME,
     SETTINGS_CAT_SYSTEM_TOOLS,
     SETTINGS_CAT_LOGGING,
+    SETTINGS_CAT_FAVORITES,
     SETTINGS_CAT_BACKUP_RESET,
     SETTINGS_CAT_SCAN_SAVING,
     SETTINGS_CAT_WIGLE,
@@ -1145,6 +1147,7 @@ static SettingsCategory settings_categories[] = {
     {"GPS", SETTINGS_CAT_GPS, SETTINGS_ROOT_DATA_TOOLS, false, NULL},
     {"Saving", SETTINGS_CAT_SCAN_SAVING, SETTINGS_ROOT_DATA_TOOLS, false, NULL},
     {"Lock Screen", SETTINGS_CAT_LOCKSCREEN, SETTINGS_ROOT_SECURITY, false, NULL},
+    {"Favorites", SETTINGS_CAT_FAVORITES, SETTINGS_ROOT_SECURITY, false, NULL},
     {"Date & Time", SETTINGS_CAT_DATE_TIME, SETTINGS_ROOT_SYSTEM, false, NULL},
     {"Power", SETTINGS_CAT_POWER, SETTINGS_ROOT_SYSTEM, false, NULL},
     {"Setup", SETTINGS_CAT_SYSTEM_TOOLS, SETTINGS_ROOT_SYSTEM, false, NULL},
@@ -1806,6 +1809,8 @@ static SettingsItem settings_items[] = {
     {"Invert Colors", SETTING_INVERT_COLORS, bool_options, 2, 0, SETTINGS_CAT_DISPLAY, false, NULL, SETTING_WIDGET_TOGGLE},
     {"Sun Mode", SETTING_SUN_MODE, bool_options, 2, 0, SETTINGS_CAT_DISPLAY, false, NULL, SETTING_WIDGET_TOGGLE},
     {"Log Level", SETTING_LOG_LEVEL, log_level_options, 6, ESP_LOG_WARN, SETTINGS_CAT_LOGGING, false, NULL, SETTING_WIDGET_VALUE_CYCLE},
+    {"Open Without PIN", SETTING_FAVORITES_BYPASS, bool_options, 2, 0, SETTINGS_CAT_FAVORITES, false, NULL, SETTING_WIDGET_TOGGLE},
+    {"Manage Favorites", SETTING_MANAGE_FAVORITES, action_options, 1, 0, SETTINGS_CAT_FAVORITES, false, NULL, SETTING_WIDGET_VALUE_CYCLE},
     {"Terminal Font", SETTING_TERMINAL_FONT_SIZE, font_size_options, 3, 1, SETTINGS_CAT_DISPLAY, false, NULL, SETTING_WIDGET_VALUE_CYCLE},
 
     {"Menu Theme", SETTING_MENU_THEME, theme_options, 17, 0, SETTINGS_CAT_THEME_ASSETS, false, NULL, SETTING_WIDGET_VALUE_CYCLE},
@@ -4322,6 +4327,12 @@ static void load_current_settings_values(void) {
             case SETTING_LOG_LEVEL:
                 settings_items[i].current_value = settings_get_log_level(&G_Settings);
                 break;
+            case SETTING_FAVORITES_BYPASS:
+                settings_items[i].current_value = settings_get_favorites_bypass(&G_Settings) ? 1 : 0;
+                break;
+            case SETTING_MANAGE_FAVORITES:
+                settings_items[i].current_value = 0;
+                break;
             case SETTING_WEB_AUTH:
                 settings_items[i].current_value = settings_get_web_auth_enabled(&G_Settings) ? 1 : 0;
                 break;
@@ -4621,6 +4632,12 @@ static void apply_setting_change(int setting_index, int new_value) {
             settings_set_log_level(&G_Settings, (uint8_t)new_value);
             esp_log_level_set("*", (esp_log_level_t)new_value);
             break;
+        case SETTING_FAVORITES_BYPASS:
+            settings_set_favorites_bypass(&G_Settings, new_value == 1);
+            break;
+        case SETTING_MANAGE_FAVORITES:
+            display_manager_switch_view(&favorites_manager_view);
+            return;
         case SETTING_WEB_AUTH:
             settings_set_web_auth_enabled(&G_Settings, new_value == 1);
             break;